Move 'Authentication' to Security namespace
