+++ index.php
@@ -38,6 +38,6 @@ $a->runFrontend(
 	$dice->create(\Friendica\App\Module::class),
 	$dice->create(\Friendica\App\Router::class),
 	$dice->create(\Friendica\Core\PConfig\IPConfig::class),
-	$dice->create(\Friendica\App\Authentication::class),
+	$dice->create(\Friendica\Security\Authentication::class),
 	$dice->create(\Friendica\App\Page::class)
 );
